python_server/pyproject.toml
[project]
name = "python-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "pillow>=11.3.0",
    "tensorflow>=2.20.0",
]
